Type PSCI_SYSTEM_OFF as usize
It is a function ID passed in x0, and everything else in board that lands in a
register is already usize.

+++ src/board.rs
@@ -6,4 +6,4 @@ pub const GICC_BASE: usize = 0x0801_0000;
 pub const UART_INTID: u32 = 33;
 
 // PSCI (Power State Coordination Interface) function ID for SYSTEM_OFF.
-pub const PSCI_SYSTEM_OFF: u64 = 0x8400_0008;
+pub const PSCI_SYSTEM_OFF: usize = 0x8400_0008;
